requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-Migrate>=4.0.5
Werkzeug>=3.0.0
pandas>=2.1.0
scikit-learn>=1.3.0
shap>=0.43.0
numpy>=1.26.0
python-dotenv>=1.0.0
email-validator>=2.1.0
requests>=2.31.0
Flask-Mail>=0.9.1
